+++ pandas/tests/groupby/test_quantile.py
@@ -215,122 +215,3 @@ def test_groupby_quantile_nullable_array(values, q):
 
     expected = pd.Series(true_quantiles * 2, index=idx, name="b")
     tm.assert_series_equal(result, expected)
-
-
-# pipe
-# --------------------------------
-
-
-def test_pipe():
-    # Test the pipe method of DataFrameGroupBy.
-    # Issue #17871
-
-    random_state = np.random.RandomState(1234567890)
-
-    df = DataFrame(
-        {
-            "A": ["foo", "bar", "foo", "bar", "foo", "bar", "foo", "foo"],
-            "B": random_state.randn(8),
-            "C": random_state.randn(8),
-        }
-    )
-
-    def f(dfgb):
-        return dfgb.B.max() - dfgb.C.min().min()
-
-    def square(srs):
-        return srs ** 2
-
-    # Note that the transformations are
-    # GroupBy -> Series
-    # Series -> Series
-    # This then chains the GroupBy.pipe and the
-    # NDFrame.pipe methods
-    result = df.groupby("A").pipe(f).pipe(square)
-
-    index = Index(["bar", "foo"], dtype="object", name="A")
-    expected = pd.Series([8.99110003361, 8.17516964785], name="B", index=index)
-
-    tm.assert_series_equal(expected, result)
-
-
-def test_pipe_args():
-    # Test passing args to the pipe method of DataFrameGroupBy.
-    # Issue #17871
-
-    df = pd.DataFrame(
-        {
-            "group": ["A", "A", "B", "B", "C"],
-            "x": [1.0, 2.0, 3.0, 2.0, 5.0],
-            "y": [10.0, 100.0, 1000.0, -100.0, -1000.0],
-        }
-    )
-
-    def f(dfgb, arg1):
-        return dfgb.filter(lambda grp: grp.y.mean() > arg1, dropna=False).groupby(
-            dfgb.grouper
-        )
-
-    def g(dfgb, arg2):
-        return dfgb.sum() / dfgb.sum().sum() + arg2
-
-    def h(df, arg3):
-        return df.x + df.y - arg3
-
-    result = df.groupby("group").pipe(f, 0).pipe(g, 10).pipe(h, 100)
-
-    # Assert the results here
-    index = pd.Index(["A", "B", "C"], name="group")
-    expected = pd.Series([-79.5160891089, -78.4839108911, -80], index=index)
-
-    tm.assert_series_equal(expected, result)
-
-    # test SeriesGroupby.pipe
-    ser = pd.Series([1, 1, 2, 2, 3, 3])
-    result = ser.groupby(ser).pipe(lambda grp: grp.sum() * grp.count())
-
-    expected = pd.Series([4, 8, 12], index=pd.Int64Index([1, 2, 3]))
-
-    tm.assert_series_equal(result, expected)
-
-
-def test_groupby_mean_no_overflow():
-    # Regression test for (#22487)
-    df = pd.DataFrame(
-        {
-            "user": ["A", "A", "A", "A", "A"],
-            "connections": [4970, 4749, 4719, 4704, 18446744073699999744],
-        }
-    )
-    assert df.groupby("user")["connections"].mean()["A"] == 3689348814740003840
-
-
-@pytest.mark.parametrize(
-    "values",
-    [
-        {
-            "a": [1, 1, 1, 2, 2, 2, 3, 3, 3],
-            "b": [1, pd.NA, 2, 1, pd.NA, 2, 1, pd.NA, 2],
-        },
-        {"a": [1, 1, 2, 2, 3, 3], "b": [1, 2, 1, 2, 1, 2]},
-    ],
-)
-@pytest.mark.parametrize("function", ["mean", "median", "var"])
-def test_apply_to_nullable_integer_returns_float(values, function):
-    # https://github.com/pandas-dev/pandas/issues/32219
-    output = 0.5 if function == "var" else 1.5
-    arr = np.array([output] * 3, dtype=float)
-    idx = pd.Index([1, 2, 3], dtype=object, name="a")
-    expected = pd.DataFrame({"b": arr}, index=idx)
-
-    groups = pd.DataFrame(values, dtype="Int64").groupby("a")
-
-    result = getattr(groups, function)()
-    tm.assert_frame_equal(result, expected)
-
-    result = groups.agg(function)
-    tm.assert_frame_equal(result, expected)
-
-    result = groups.agg([function])
-    expected.columns = MultiIndex.from_tuples([("b", function)])
-    tm.assert_frame_equal(result, expected)
